pyproject.toml
[project]
name = "convergeai-ai-agent-job"
version = "0.1.0"
description = "AI solution for job search and personalized cover letter generation in Danish and English."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=0.6.3",
    "scrapy>=2.12.0",
    "sentence-transformers>=3.4.1",
    "accelerate==1.3.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio>=2.6.0",
    "transformers==4.48.2",
    "peft==0.13.2",
    "pillow>=10.4",
    "scipy>=1.15.2",
    "backoff==2.2.1",
    "streamlit>=1.43.2",
    "soundfile>=0.13.1",
    "flash_attn==2.7.4.post1",
    "huggingface-hub>=0.29.3",
    "fastapi>=0.115.11",
    "uvicorn>=0.34.0",
    "jinja2>=3.1.6",
    "chromadbx>=0.1.0",
    "pdfplumber>=0.11.5",
    "python-docx>=1.1.2",
    "nest-asyncio>=1.6.0",
    "crawl4ai[sync]>=0.5.0.post4",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" },
]
torchvision = [
    { index = "pytorch-cu124" },
]
torchaudio = [
    { index = "pytorch-cu124" },
]
flash-attn = { url = "https://github.com/kingbri1/flash-attention/releases/download/v2.7.4.post1/flash_attn-2.7.4.post1+cu124torch2.6.0cxx11abiFALSE-cp312-cp312-win_amd64.whl" }


[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

